+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## [0.1.1] - 2021-08-28
+
+### Changed
+
+- Rename `mode` options `:replace` and `:delete` to `{:action, :replace}` and
+  `{:action, :delete}`.
+
 ## [0.1.0] - 2021-08-28
 
 initial release